# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for Laravel application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Get users list",
                "description": "Display a listing of the resource.",
                "operationId": "2979f0297ab881ad0b8b86ee613f077b",
                "responses": {
                    "200": {
                        "description": "Users list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumBearer": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Create a user",
                "description": "Store a newly created resource in storage.",
                "operationId": "7601d6cfa9768b3857d0e58d45eb618b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe",
                                        "maxLength": 255,
                                        "minLength": 3
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123",
                                        "minLength": 8
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "editor",
                                        "enum": [
                                            "super-admin",
                                            "admin",
                                            "editor",
                                            "user"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumBearer": []
                    },
                    {
                        "sanctumCookie": [],
                        "xsrfToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResource": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "roles"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "editor"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctumBearer": {
                "type": "http",
                "description": "Use a Laravel Sanctum personal access token in the Authorization header.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            },
            "sanctumCookie": {
                "type": "apiKey",
                "description": "Laravel Sanctum SPA session cookie.",
                "name": "laravel_session",
                "in": "cookie"
            },
            "xsrfToken": {
                "type": "apiKey",
                "description": "CSRF token header required for stateful Sanctum SPA requests.",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}